+++ slipp/src/main/java/slipp/support/context/NotFoundBeanException.java
@@ -1,4 +1,4 @@
-package slipp;
+package slipp.support.context;
 
 public class NotFoundBeanException extends RuntimeException {
     private static final String MESSAGE = "해당 Bean이 존재하지 않습니다. : ";
